fix: 닉네임 중복 처리

+++ src/main/java/com/ocupid/server/controller/AuthController.java
@@ -53,10 +53,12 @@ public class AuthController {
             emailVerificationService.getEmailVerificationInfo(request.getEmail())
                 .orElseThrow(() -> new NotVerifiedEmailException("인증되지 않은 이메일입니다."));
 
-        // TODO: 이메일, 닉네임 중복 체크
-
         if (!emailVerification.getIsChecked()) {
             throw new NotVerifiedEmailException("인증되지 않은 이메일입니다.");
+        }
+
+        if (!userService.isExist(request.getEmail())) {
+            throw new AlreadyExistResourceException("중복된 닉네임입니다.");
         }
 
         if (!emailVerificationService.deleteVerificationInfo(emailVerification.getId())) {
